pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "client-portal"
version = "0.1.0"
description = "Repeatable client analytics portal: upload → analyze → dashboard → PowerPoint"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "pyarrow>=14.0",
  "pydantic>=2.0",
  "streamlit>=1.30",
  "plotly>=5.18",
  "kaleido>=0.2.1",
  "python-pptx>=0.6.23",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5",
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "pre-commit>=3.7",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
